Split Provider helpers into focused extensions
Refactored `Provider.swift` by moving error and fetch helper logic into `Provider+Error.swift` and `Provider+Fetch.swift`, keeping the core protocol definition lean. This also adds `fetchIfFailed(in:)`, allowing providers to retry only when the previous fetch ended in an error, while preserving existing cancellation and retry behavior.
